chore(__tests__/resources/**) test folderified
