+++ __tests__/resources/patients/index.js
@@ -1,7 +1,7 @@
 // @flow
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import * as sdk from '../../src';
+import * as sdk from '../../../src/index';
 
 jest.setTimeout(10000);
 let instance = {};
